% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_mcSimulation.R
\name{cast_mcSimulation}
\alias{cast_mcSimulation}
\title{Extract the simulated output from \code{\link[decisionSupport]{mcSimulation}} object}
\usage{
cast_mcSimulation(x)
}
\arguments{
\item{x}{\code{\link[decisionSupport]{mcSimulation}} or a list of \code{\link[decisionSupport]{mcSimulation}}
from which to extract the simulated output.}
}
\value{
If x is a list, then a list of simulated output is returned. Otherwise, a dataframe is returned.
}
\description{
Take an object of class \code{\link[decisionSupport]{mcSimulation}} and return the simulated values
as specified in the model function.
}
\author{
Issoufou Liman
}
