% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dens_mcsimilation_data.R
\name{get_dens_mcsimilation_data}
\alias{get_dens_mcsimilation_data}
\title{Format \code{\link[decisionSupport]{mcSimulation}} data for ggplot}
\usage{
get_dens_mcsimilation_data(x, colorQuantile = c("GRAY48", "YELLOW",
  "ORANGE", "DARK GREEN", "ORANGE1", "YELLOW1", "GRAY49"),
  colorProbability = c(1, 0.95, 0.75, 0.55, 0.45, 0.25, 0.05))
}
\arguments{
\item{x}{\code{\link[decisionSupport]{mcSimulation}} or a list of \code{\link[decisionSupport]{mcSimulation}}
from which to extract the simulated output.}

\item{colorQuantile}{\code{character vector}: encoding the colors of the 
quantiles defined in argument \code{colorProbability}.}

\item{colorProbability}{\code{numeric vector}: defines the quantiles that 
shall be distinguished by the colors chosen in argument 
\code{colorQuantile}. Must be of the same length as \code{colorQuantile}.}
}
\value{
If x is a list, then a list of simulated output is returned. Otherwise, a dataframe is returned.
}
\description{
Take an object of class \code{\link[decisionSupport]{mcSimulation}} or a list of these and return the simulated values
such that, when converted into molten data frame, can be easily understood by ggplot.
}
\details{
The specification are made for \code{\link[ggplot2]{geom_density}}.
}
\seealso{
\code{\link[decisionSupportExtra]{get_dens_mcsimilation_data}} \code{\link[decisionSupport]{mcSimulation}}
}
\author{
Issoufou Liman
}
