% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbuild_descdist_data.R
\name{ggbuild_descdist_data}
\alias{ggbuild_descdist_data}
\title{Format \code{\link[fitdistrplus]{descdist}} data to ggplot2}
\usage{
ggbuild_descdist_data(
  data,
  discrete = FALSE,
  boot = NULL,
  method = "unbiased",
  graph = TRUE,
  title = "Cullen and Frey graph",
  subtitle = NULL,
  obs.col = "darkblue",
  boot.col = "orange"
)
}
\arguments{
\item{data}{ A numeric vector.  }

\item{discrete}{ If \code{TRUE}, the distribution is considered as discrete.  }

\item{boot}{ If not \code{NULL}, \code{boot} values of skewness and kurtosis are plotted from bootstrap
 samples of \code{data}. \code{boot} must be fixed in this case to an integer above 10. }

\item{method}{ "unbiased" for unbiased estimated values of statistics or "sample" for sample values. }

\item{graph}{ If \code{FALSE}, the skewness-kurtosis graph is not plotted. }

\item{title, subtitle}{Title and Subtitle}

\item{obs.col}{ Color used for the observed point on the skewness-kurtosis graph. }

\item{boot.col}{ Color used for bootstrap sample of points on the skewness-kurtosis graph. }
}
\description{
Generate a list of \code{\link[fitdistrplus]{descdist}} outputs to be understood by ggplot2.
}
\details{
see \code{\link[fitdistrplus]{descdist}}.
}
\examples{
ggbuild_descdist_data(rbeta(100, shape1 = 0.05, shape2 = 1), boot = 500, obs.col = 'blue',
boot.col = 'yellow')
}
\references{
Marie Laure Delignette-Muller, Christophe Dutang (2015). fitdistrplus: An R Package for Fitting Distributions. Journal of Statistical Software, 64(4), 1-34. http://www.jstatsoft.org/v64/i04/.
}
\seealso{
\code{\link[fitdistrplus]{descdist}}.
}
\author{
Issoufou Liman
}
