% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_descdist_bn.R
\name{ggplot_descdist_bn}
\alias{ggplot_descdist_bn}
\title{ggplot the posterior distribution of a Bayesian network node as in \code{\link[fitdistrplus]{descdist}}}
\usage{
ggplot_descdist_bn(bn, node, boot = 1000, obs.col = "darkblue",
  boot.col = "orange", title = "Cullen and Frey graph",
  subtitle = node, xlab = "square of skewness", ylab = "kurtosis",
  obs_geom_size = 4, boot_geom_size = 0.02, dist_geom_pts_size = 5,
  dist_geom_line_size = 0.6, axis_text_size = 12,
  axis_title_size = 12, plot_title_size = 20,
  plot_subtitle_size = 17, strip_text_size = 18,
  legend_text_size = 12, evidence = NULL, n_generation = NULL,
  include_relatives = TRUE, n_run = 1000)
}
\arguments{
\item{bn}{an object of class bn.fit.}

\item{node}{character string, the label of the node which conditional distribution is of interest.}

\item{boot}{ If not \code{NULL}, \code{boot} values of skewness and kurtosis are plotted from bootstrap
 samples of \code{data}. \code{boot} must be fixed in this case to an integer above 10. }

\item{obs.col}{ Color used for the observed point on the skewness-kurtosis graph. }

\item{boot.col}{ Color used for bootstrap sample of points on the skewness-kurtosis graph. }

\item{title}{Title and Subtitle}

\item{subtitle}{Title and Subtitle}

\item{xlab}{These are respectively x and y labels.}

\item{ylab}{These are respectively x and y labels.}

\item{obs_geom_size}{The size of the geom_point to be used for the empirical distributoion (default to 4), bootstrapping (default to 0.02), theoritical distribution (default to 5), respectively.}

\item{boot_geom_size}{The size of the geom_point to be used for the empirical distributoion (default to 4), bootstrapping (default to 0.02), theoritical distribution (default to 5), respectively.}

\item{dist_geom_pts_size}{The size of the geom_point to be used for the empirical distributoion (default to 4), bootstrapping (default to 0.02), theoritical distribution (default to 5), respectively.}

\item{dist_geom_line_size}{The size of the geom_line to be used for the empirical distributoion. The default is 0.6.}

\item{axis_text_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{axis_title_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{plot_title_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{plot_subtitle_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{strip_text_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{legend_text_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{evidence}{a name value pair:a named character vector which values are node states and names are node names.}

\item{n_generation}{how far to go in the network topology for building the conditionning specification for the query?}

\item{include_relatives}{logical Should parents or ancestors, depending on the the argument n_generation, should be included in the query?
If TRUE, the default, these will be internally involved in constructing the evidence argument.}

\item{n_run}{integer specifying the number of of model run. Default is 1000.}
}
\description{
Generate compact plot of Bayesian network node states following the \code{\link[fitdistrplus]{descdist}} function as customizable ggplot.
}
\details{
see \code{\link[fitdistrplus]{descdist}}.
}
\examples{
library (gRain)
library(bnlearn)
## setting a bayesian network with gRain
Soil_type <- cptable (~Soil_type, values = c(0.05, 0.55, 0.4),
levels = c('Sandy', 'Loamy', 'Clayey'))
Manure_application <- cptable(~Manure_application, values = c(0.3, 0.7),
levels = c('FALSE', 'TRUE'))
Soil_water_holding_capacity_tmp <- make_gRain_CPT(
 parent_effects = list(c(0, 2.5, 3), c(0, 2)),
 parent_weights = c(2,1),
 b = 3,
 child_prior = c(0.2,0.5,0.3),
 child_states = c('Low', 'Medium', 'High'),
 parent_states = list(c('Sandy', 'Loamy', 'Clayey'), c('FALSE', 'TRUE'))
)
Soil_water_holding_capacity_values <- Soil_water_holding_capacity_tmp$values
Soil_water_holding_capacity_levels <- Soil_water_holding_capacity_tmp$levels
Soil_water_holding_capacity <- cptable (
~Soil_water_holding_capacity|Soil_type:Manure_application,
values = Soil_water_holding_capacity_values,
levels = Soil_water_holding_capacity_levels)
## Compile conditional probability tables
network <- compileCPT(list(Soil_type, Manure_application, Soil_water_holding_capacity))
## Graphical Independence Network ####
network <- grain(network)
## Use grain object (gRain package)
ggplot_descdist_bn (bn = network, node = 'Soil_water_holding_capacity')

## converting the grain bayesian network to bn.fit
network_bn_fit <- as.bn.fit(network)
## Use bn.fit object (bnlearn package)
ggplot_descdist_bn (bn = network, node = 'Soil_water_holding_capacity')
}
\references{
Marie Laure Delignette-Muller, Christophe Dutang (2015). fitdistrplus: An R Package for Fitting Distributions. Journal of Statistical Software, 64(4), 1-34. http://www.jstatsoft.org/v64/i04/.
}
\seealso{
\code{\link[fitdistrplus]{descdist}}.
}
\author{
Issoufou Liman
}
