% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dens_mcsimilation_data.R
\name{ggplot_mc_dens}
\alias{ggplot_mc_dens}
\alias{ggplot_mc_dens.mcSimulation}
\alias{ggplot_mc_dens.list}
\title{Generic \code{\link[decisionSupport]{mcSimulation}} density plotting with ggplot}
\usage{
ggplot_mc_dens(x, ...)

\method{ggplot_mc_dens}{mcSimulation}(x, ...)

\method{ggplot_mc_dens}{list}(x, ...)
}
\arguments{
\item{x}{\code{\link[decisionSupport]{mcSimulation}} or a list of \code{\link[decisionSupport]{mcSimulation}}}

\item{...}{Additional Arguments such as colorQuantile and colorProbability passed to get_dens_mcsimilation_data}
}
\value{
A ggplot object
}
\description{
generate density for \code{\link[decisionSupport]{mcSimulation}} using \code{\link[ggplot2]{geom_density}}
}
\seealso{
\code{\link[decisionSupportExtra]{ggplot_mc_hist}} \code{\link[decisionSupportExtra]{get_dens_mcsimilation_data}} \code{\link[decisionSupport]{mcSimulation}}
}
\author{
Issoufou Liman
}
