% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hist_mcsimilation_data.R
\name{ggplot_mc_hist}
\alias{ggplot_mc_hist}
\alias{ggplot_mc_hist.mcSimulation}
\alias{ggplot_mc_hist.list}
\title{Generic mcSimulation histogram plotting with ggplot}
\usage{
ggplot_mc_hist(x, ...)

\method{ggplot_mc_hist}{mcSimulation}(x, ...)

\method{ggplot_mc_hist}{list}(x, ...)
}
\arguments{
\item{x}{\code{\link[decisionSupport]{mcSimulation}} or a list of \code{\link[decisionSupport]{mcSimulation}}}

\item{...}{Additional Arguments such as breaks, colorQuantile and colorProbability passed to get_hist_mcsimilation_data}
}
\value{
A ggplot object
}
\description{
generate histogram for \code{\link[decisionSupport]{mcSimulation}} using \code{\link[ggplot2]{geom_col}}
}
\seealso{
\code{\link[decisionSupportExtra]{ggplot_mc_dens}} \code{\link[decisionSupportExtra]{get_hist_mcsimilation_data}} \code{\link[decisionSupport]{mcSimulation}}
}
\author{
Issoufou Liman
}
