% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cpdist.R
\name{sample_cpdist}
\alias{sample_cpdist}
\title{Take a bayesian network, a node of interst, and a return value type option and returns either the
probability/sampled data (depending on the op argument specified) of the different  states of the
node of interst given the different conbinaition of its and its parents states.}
\usage{
sample_cpdist(bn, node, op = c("sampler", "proba"), evidence = NULL,
  n_generation = NULL)
}
\arguments{
\item{bn}{an object of class bn.fit.}

\item{node}{a vector of character strings, the labels of the nodes whose conditional distribution we are interested in.}

\item{op}{a vector of character strings, the type of returned value: either probabilities or raw data sampled from the posterior distribution}

\item{evidence}{a name value pair:a named character vector which values are node states and names are node names.}

\item{n_generation}{how far to go in the network topology for building the conditionning specification for the query?}
}
\value{
a matrix containing either the probabilities or raw data sampled from the posterior distribution
}
\description{
Take a bayesian network, a node of interst, and a return value type option and returns either the
probability/sampled data (depending on the op argument specified) of the different  states of the
node of interst given the different conbinaition of its and its parents states.
}
\details{
the rownames are specify the conditionning used for each query.
}
