% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub-estimate-method.R
\name{[.estimate}
\alias{[.estimate}
\title{Extract or replace subsets of \code{\link[decisionSupport]{estimate}}}
\usage{
\method{[}{estimate}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{estimate object}

\item{i, j}{elements to extract or replace by index or dimnames.}

\item{drop}{logical. If TRUE the result is coerced to the lowest possible dimension. The default is to drop if only one column is left, but not to drop if only one row is left.}
}
\description{
subset operator for estimate class from decisionSupport package.
}
\details{
Attempt to drop any of the required estimate columns (i.e distribution, min, max )
is silently ignored since estimate object may not make sense without them.
}
\examples{
## This example is taken fom decisionSupport
library (decisionSupport)
# Create an estimate from text (with correlated components):
estimateTextMarg<-'variable,  distribution, lower, upper
revenue1,  posnorm,      100,   1000
revenue2,  posnorm,      50,    2000
costs1,    posnorm,      50,    2000
costs2,    posnorm,      100,   1000'
estimateTextCor<-',         revenue1, costs2
revenue1,        1,   -0.3
costs2,       -0.3,      1'
estimateCor <- as.estimate (read.csv (header=TRUE, text=estimateTextMarg,
strip.white = TRUE, stringsAsFactors = FALSE),
correlation_matrix = data.matrix (read.csv (text = estimateTextCor,
row.names = 1, strip.white = TRUE)))
## extracting the first row.
estimateCor [1, ]
## extracting all the rows keep everything intact.
estimateCor [1:4, ]
## Trying to drop mandatory column is ignored.
estimateCor [, 1]
}
\author{
Issoufou Liman
}
