% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dens_mcsimilation_data.R,
%   R/get_hist_mcsimilation_data.R
\name{get_dens_mcsimilation_data}
\alias{get_dens_mcsimilation_data}
\alias{get_hist_mcsimilation_data}
\title{Compute kernel density and histogram data, from \code{\link[decisionSupport]{mcSimulation}} object, for \code{\link[ggplot2]{geom_bar}} and \code{\link[ggplot2]{geom_density}}}
\usage{
get_dens_mcsimilation_data(x, colorQuantile = c("GRAY48", "YELLOW",
  "ORANGE", "DARK GREEN", "ORANGE1", "YELLOW1", "GRAY49"),
  colorProbability = c(1, 0.95, 0.75, 0.55, 0.45, 0.25, 0.05))

get_hist_mcsimilation_data(x, breaks = 100, colorQuantile = c("GRAY48",
  "YELLOW", "ORANGE", "DARK GREEN", "ORANGE1", "YELLOW1", "GRAY49"),
  colorProbability = c(1, 0.95, 0.75, 0.55, 0.45, 0.25, 0.05))
}
\arguments{
\item{x}{\code{\link[decisionSupport]{mcSimulation}} or a list of \code{\link[decisionSupport]{mcSimulation}}}

\item{colorQuantile}{\code{character vector}: encoding the colors of the 
quantiles defined in argument \code{colorProbability}.}

\item{colorProbability}{\code{numeric vector}: defines the quantiles that 
shall be distinguished by the colors chosen in argument 
\code{colorQuantile}. Must be of the same length as \code{colorQuantile}.}

\item{breaks}{one of:
    \itemize{
      \item a vector giving the breakpoints between histogram cells,
      \item a function to compute the vector of breakpoints,
      \item a single number giving the number of cells for the histogram,
      \item a character string naming an algorithm to compute the
      number of cells (see \sQuote{Details}),
      \item a function to compute the number of cells.
    }
    In the last three cases the number is a suggestion only; as the
    breakpoints will be set to \code{\link{pretty}} values, the number
    is limited to \code{1e6} (with a warning if it was larger).  If
    \code{breaks} is a function, the \code{x} vector is supplied to it
    as the only argument (and the number of breaks is only limited by
    the amount of available memory).
  }
}
\value{
\itemize{
\item{list: }{If x is a list object.}
\item{dataframe: }{If x is an mcSimulation object.}
}
}
\description{
Take an object of class \code{\link[decisionSupport]{mcSimulation}} or a list of these and return a ggplot object
or the simulated values such that, when converted into molten data frame, can be easily
understood by ggplot.
}
\section{get_hist_mcsimilation_data}{

get_hist_mcsimilation_data, which specification are made for \code{\link[ggplot2]{geom_bar}},
takes care of extracting the simulated values as histogram
from the \code{\link[decisionSupport]{mcSimulation}} object.
}

\section{get_dens_mcsimilation_data}{

get_dens_mcsimilation_data, which specification are made for \code{\link[ggplot2]{geom_density}},
takes care of extracting the simulated values as kernel density
from the \code{\link[decisionSupport]{mcSimulation}} object.
}

\section{ggplot_mc_hist and ggplot_mc_dens}{

These are generics for which methods are defined for \code{\link[decisionSupport]{mcSimulation}}
object or a list of these.Internally, these methods rely on get_hist_mcsimilation_data and get_dens_mcsimilation_data
respectively to generate the data required for \code{\link[ggplot2]{geom_bar}} and \code{\link[ggplot2]{geom_density}}
and hand over these data to ggplot.
}

\seealso{
\code{\link[decisionSupportExtra]{ggplot_mc_hist}} \code{\link[decisionSupportExtra]{get_dens_mcsimilation_data}} \code{\link[decisionSupport]{mcSimulation}}
}
\author{
Issoufou Liman
}
