% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_plotdist.R
\name{gg_plotdist}
\alias{gg_plotdist}
\title{ggplot empirical and theoretical distributions for non-censored data following the \code{\link[fitdistrplus]{descdist}} function.}
\usage{
gg_plotdist(
  fitted,
  title = NULL,
  hist_bar_size = 0.3,
  dist_geom_pts_size = 0.03,
  dist_geom_smooth_size = 0.3,
  dist_geom_abline_size = 0.3,
  axis_text_size = 9,
  axis_title_size = 9,
  plot_title_size = 10,
  plot_subtitle_size = 8,
  strip_text_size = 9
)
}
\arguments{
\item{fitted}{A list of \code{\link[fitdistrplus]{fitdist}} object}

\item{title}{character string to be used as plot title}

\item{hist_bar_size}{numeric size of histogram bars}

\item{dist_geom_pts_size, dist_geom_smooth_size, dist_geom_abline_size}{numeric. The size of the geom_point, smoothing line and abline, respectively.}

\item{axis_text_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{axis_title_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{plot_title_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{plot_subtitle_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}

\item{strip_text_size}{= 12,
Text size respectively corresponding to axis text (default to 12), axis title (default to 12), plot title (default to 20), subtitle (default to 17), strip text (default to 18), and legend (default to 12).}
}
\description{
Generate compact plot of Bayesian network node states following the \code{\link[fitdistrplus]{descdist}} function as customizable ggplot.
}
\details{
see \code{\link[fitdistrplus]{plotdist}}.
}
\examples{
#' library (gRain)
#' library(bnlearn)
#' ## setting a bayesian network with gRain
#' Soil_type <- cptable (~Soil_type, values = c(0.05, 0.55, 0.4),
#' levels = c('Sandy', 'Loamy', 'Clayey'))
#' Manure_application <- cptable(~Manure_application, values = c(0.3, 0.7),
#' levels = c('FALSE', 'TRUE'))
#' Soil_water_holding_capacity_tmp <- make_gRain_CPT(
#'  parent_effects = list(c(0, 2.5, 3), c(0, 2)),
#'  parent_weights = c(2,1),
#'  b = 3,
#'  child_prior = c(0.2,0.5,0.3),
#'  child_states = c('Low', 'Medium', 'High'),
#'  parent_states = list(c('Sandy', 'Loamy', 'Clayey'), c('FALSE', 'TRUE'))
#' )
#' Soil_water_holding_capacity_values <- Soil_water_holding_capacity_tmp$values
#' Soil_water_holding_capacity_levels <- Soil_water_holding_capacity_tmp$levels
#' Soil_water_holding_capacity <- cptable (
#' ~Soil_water_holding_capacity|Soil_type:Manure_application,
#' values = Soil_water_holding_capacity_values,
#' levels = Soil_water_holding_capacity_levels)
#' ## Compile conditional probability tables
#' network <- compileCPT(list(Soil_type, Manure_application, Soil_water_holding_capacity))
#' ## Graphical Independence Network ####
#' network <- grain(network)
#' ## Use grain object (gRain package)
#' fitted <- fit_node_states_distr (bn = network,
#' node = "Soil_water_holding_capacity", gof="KS")
#' gg_plotdist(fitted, title = "Soil_water_holding_capacity")

#' ## converting the grain bayesian network to bn.fit
#' network_bn_fit <- as.bn.fit(network)
#' ## Use bn.fit object (bnlearn package)
#' fitted <- fit_node_states_distr (bn = network_bn_fit,
#' node = "Soil_water_holding_capacity", distr = c("beta", "beta", "beta"))
#' gg_plotdist(fitted, title = "Soil_water_holding_capacity")
}
\references{
Marie Laure Delignette-Muller, Christophe Dutang (2015). fitdistrplus: An R Package for Fitting Distributions. Journal of Statistical Software, 64(4), 1-34. http://www.jstatsoft.org/v64/i04/.
}
\seealso{
\code{\link[fitdistrplus]{plotdist}}.
}
\author{
Issoufou Liman
}
